3rd part of FS#12176. Gain setting migrated to fixed point for libgme.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30277 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libgme/ay_emu.h
@@ -66,7 +66,7 @@ struct Ay_Emu {
 	int voice_count;
 	int mute_mask_;
 	int tempo;
-	double gain;
+	int gain;
 	
 	long sample_rate;
 	
@@ -152,7 +152,7 @@ void Sound_mute_voices( struct Ay_Emu* this, int mask );
 
 // Change overall output amplitude, where 1.0 results in minimal clamping.
 // Must be called before set_sample_rate().
-static inline void Sound_set_gain( struct Ay_Emu* this, double g )
+static inline void Sound_set_gain( struct Ay_Emu* this, int g )
 {
 	assert( !this->sample_rate ); // you must set gain before setting sample rate
 	this->gain = g;
