+++ apps/codecs/libgme/hes_emu.h
@@ -70,7 +70,7 @@ struct Hes_Emu {
 	unsigned buf_changed_count;
 	int voice_count_;
 	int tempo_;
-	double gain_;
+	int gain_;
 	
 	// track-specific
 	byte track_count;
@@ -179,7 +179,7 @@ void Sound_mute_voices( struct Hes_Emu* this, int mask );
 
 // Change overall output amplitude, where 1.0 results in minimal clamping.
 // Must be called before set_sample_rate().
-static inline void Sound_set_gain( struct Hes_Emu* this, double g )
+static inline void Sound_set_gain( struct Hes_Emu* this, int g )
 {
 	assert( !this->sample_rate_ ); // you must set gain before setting sample rate
 	this->gain_ = g;
